Restructured /RequestDetails/ files into multiple folders for better organization
