+++ src/components/Firestore/Requests/RequestDetails/Header/Actions.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-@import './Header_variables.scss';
+@import './variables.scss';
 
 // Actions
 $actions-width-threshold: 850px;
