+++ src/components/Firestore/Requests/RequestDetails/Header/index.scss
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-@import '../index.scss';
-@import './Header_variables.scss';
+@import '../../../Requests/index.scss';
+@import './variables.scss';
 
 // Header return button
 $header-return-padding-right: 12px;
