+++ src/components/Firestore/Requests/RequestDetails/index.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-@import './Header_variables.scss';
+@import './Header/variables.scss';
 @import '../../Requests/variables.scss';
 
 // Details content
