+++ src/components/Firestore/index.scss
@@ -17,7 +17,7 @@
 @import '../common/color';
 @import '../common/utils';
 @import '../App/variables';
-@import './variables.scss';
+@import '../Firestore/variables';
 
 .Firestore {
   display: flex;
